refactor: add path alias for modified scroll-area
