+++ blueledger-webapp/app/(public)/page.tsx
@@ -1,32 +1,31 @@
-import Link from 'next/link';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { LandingNav } from './landing-nav';
+import { Faq } from './sections/faq/faq';
+import { Features } from './sections/features/features';
+import { Footer } from './sections/footer';
+import { Hero } from './sections/hero';
+import { Pricing } from './sections/pricing/pricing';
 
 export default function Home() {
   return (
-    <main className="flex justify-center items-center h-screen">
-      <Card className="w-[550px] mx-auto">
-        <CardHeader>
-          <CardTitle>Blue Ledger</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <p>This will be the landing page for Blue Ledger.</p>
-          <p>A platform for managing your finances.</p>
-          <br />
-          <p>
-            With the help of AI, you can easily track your income and expenses.
-          </p>
-          <br />
-          <p>
-            Go to
-            {' '}
-            <Link href="/dashboard" className="underline">
-              Dashboard
-            </Link>
-            {' '}
-            to get started.
-          </p>
-        </CardContent>
-      </Card>
-    </main>
+    <div className="dark bg-background text-foreground min-h-screen">
+      <LandingNav />
+      <header>
+        <section
+          id="hero"
+          className="landing__section landing__hero landing__grid max-w-screen-3xl mx-auto"
+        >
+          <Hero />
+        </section>
+      </header>
+      <main className="mt-16">
+        <Features />
+        <Pricing />
+        <Faq />
+      </main>
+      <div className="landing__splitter" />
+      <footer id="footer" className="landing__container">
+        <Footer />
+      </footer>
+    </div>
   );
 }
